requirements.txt
datetime
requests
python-dotenv

